全量实验代码与数据: ViT-L TAPT、四项收尾实验、管线修复与加速
代码修复与改进:
- 修复 3 份 TAPT 配置 arch 名错误与架构保真缺陷 (storage tokens/bias_mask 静默丢失)
- few-shot/benchmark 训练循环加 bf16 autocast; 冻结 backbone BN 漂移修复
- plot_confusion_matrix/save_attention 支持 PNG/PDF/SVG 三格式 (文字可编辑)
- 新增脚本: feature_cache(_tv)/eval_on_cache/supcon_adapt/convert_ckpt/
  diagnose_tapt_weights/audit_train_history/ensemble_full/stats_and_figures/summary_10shot
- train_ssl: compile 兼容、--output-dir 与 train.output_dir 自动同步
- 注册 vit_small_plus (官方 vitsplus 架构); 各实验配置对齐新产物路径

实验产物:
- TAPT (Gram/无Gram) × {S,S+,B,L} 全部 8 个 teacher 权重转换产物与收敛数据
- few-shot 全网格 120+60+40 组合; 监督基线公平协议 140+90 组合
- 消融全套 80 任务; 分辨率扫描至 672px; 集成 S+B+L (76.43%)
- 显著性检验 525 组; 实验原始数据 CSV 30 个; 两份详尽实验报告

+++ dinov3/train/train.py
@@ -108,28 +108,28 @@ def build_schedulers(cfg):
     lr = dict(
         base_value=cfg.optim["lr"],
         final_value=cfg.optim["min_lr"],
-        total_iters=cfg.optim["epochs"] * OFFICIAL_EPOCH_LENGTH,
-        warmup_iters=cfg.optim["warmup_epochs"] * OFFICIAL_EPOCH_LENGTH,
+        total_iters=int(cfg.optim["epochs"] * OFFICIAL_EPOCH_LENGTH),
+        warmup_iters=int(cfg.optim["warmup_epochs"] * OFFICIAL_EPOCH_LENGTH),
         start_warmup_value=0,
         trunc_extra=cfg.optim["schedule_trunc_extra"],
     )
     wd = dict(
         base_value=cfg.optim["weight_decay"],
         final_value=cfg.optim["weight_decay_end"],
-        total_iters=cfg.optim["epochs"] * OFFICIAL_EPOCH_LENGTH,
+        total_iters=int(cfg.optim["epochs"] * OFFICIAL_EPOCH_LENGTH),
         trunc_extra=cfg.optim["schedule_trunc_extra"],
     )
     momentum = dict(
         base_value=cfg.teacher["momentum_teacher"],
         final_value=cfg.teacher["final_momentum_teacher"],
-        total_iters=cfg.optim["epochs"] * OFFICIAL_EPOCH_LENGTH,
+        total_iters=int(cfg.optim["epochs"] * OFFICIAL_EPOCH_LENGTH),
         trunc_extra=cfg.optim["schedule_trunc_extra"],
     )
     teacher_temp = dict(
         base_value=cfg.teacher["teacher_temp"],
         final_value=cfg.teacher["teacher_temp"],
-        total_iters=cfg.teacher["warmup_teacher_temp_epochs"] * OFFICIAL_EPOCH_LENGTH,
-        warmup_iters=cfg.teacher["warmup_teacher_temp_epochs"] * OFFICIAL_EPOCH_LENGTH,
+        total_iters=int(cfg.teacher["warmup_teacher_temp_epochs"] * OFFICIAL_EPOCH_LENGTH),
+        warmup_iters=int(cfg.teacher["warmup_teacher_temp_epochs"] * OFFICIAL_EPOCH_LENGTH),
         start_warmup_value=cfg.teacher["warmup_teacher_temp"],
     )
 
@@ -139,7 +139,7 @@ def build_schedulers(cfg):
     teacher_temp_schedule = CosineScheduler(**teacher_temp)
     last_layer_lr_schedule = CosineScheduler(**lr)
 
-    last_layer_lr_schedule.schedule[: cfg.optim["freeze_last_layer_epochs"] * OFFICIAL_EPOCH_LENGTH] = (
+    last_layer_lr_schedule.schedule[: int(cfg.optim["freeze_last_layer_epochs"] * OFFICIAL_EPOCH_LENGTH)] = (
         0  # mimicking the original schedules
     )
     logger.info("Schedulers ready.")
